Split noise frame helper into multiple modules
This will make it easier for future changes
